requirements.txt
mcp
google-generativeai
python-dotenv
httpx
PyGitHub
pydantic
pydantic-settings
rich
